requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.5
python-multipart==0.0.20
pydantic==2.9.2
pydantic-settings==2.6.1
sqlmodel==0.0.22
watchdog==5.0.3
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0
pdfminer.six==20240706
python-dateutil==2.9.0.post0
